[#9]Fix: 카카오 로그인 설정 수정

+++ YAPP-BE/src/main/java/yapp/domain/member/entitiy/Member.java
@@ -28,7 +28,7 @@ public class Member extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "seq", columnDefinition = "BIGINT")
+  @Column(name = "seq", columnDefinition = "INT")
   private Long seq;
 
   @Column(name = "member_id")
